+++ modules/Accounting/Menu.php
@@ -15,12 +15,14 @@ $menu['Accounting']['admin'] = array(
 	'default' => 'Accounting/Incomes.php',
 	'Accounting/Incomes.php' => _( 'Incomes' ),
 	'Accounting/Expenses.php' => _( 'Expenses' ),
-	'Accounting/DailyTransactions.php' => _( 'Daily Transactions' ),
 	1 => _( 'Staff Payroll' ),
 	'Accounting/Salaries.php' => _( 'Salaries' ),
 	'Accounting/StaffPayments.php' => _( 'Staff Payments' ),
+	2 => _( 'Reports' ),
+	'Accounting/DailyTransactions.php' => _( 'Daily Transactions' ),
+	'Accounting/DailyTotals.php' => _( 'Daily Totals' ),
 	'Accounting/StaffBalances.php' => _( 'Staff Balances' ),
-	'Accounting/Statements.php' => _( 'Print Statements' )
+	'Accounting/Statements.php' => _( 'Print Statements' ),
 );
 
 $menu['Accounting']['teacher'] = array(
@@ -29,7 +31,8 @@ $menu['Accounting']['teacher'] = array(
 	1 => _( 'Staff Payroll' ),
 	'Accounting/Salaries.php' => _( 'Salaries' ),
 	'Accounting/StaffPayments.php' => _( 'Staff Payments' ),
-	'Accounting/Statements.php&_ROSARIO_PDF' => _( 'Print Statements' )
+	2 => _( 'Reports' ),
+	'Accounting/Statements.php&_ROSARIO_PDF' => _( 'Print Statements' ),
 );
 
 $menu['Accounting']['parent'] = array(